Title: 시력 검사 2, Time: 141ms, Memory: 11MB, Status: Wrong Answer - CodeTree

+++ 240315/시력 검사 2/eye-test-2.java
@@ -3,12 +3,12 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        float sight = sc.nextFloat();
+        double sight = sc.nextDouble();
         if (sight >= 1.0)
             System.out.println("High");
         else if (sight >= 0.5)
             System.out.println("Middle");
-        else 
+        else
             System.out.println(":Low");
 
         sc.close();
